backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.19.0
ag-ui-protocol==0.1.10
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
argcomplete==3.6.3
boto3==1.42.30
botocore==1.42.30
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cohere==5.20.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fastapi==0.128.0
fastavro==1.12.1
filelock==3.19.1
fsspec==2025.10.0
genai-prices==0.0.51
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
griffe==1.14.0
groq==1.0.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
importlib_metadata==8.7.1
invoke==2.2.1
jiter==0.12.0
jmespath==1.0.1
logfire-api==4.19.0
markdown-it-py==3.0.0
mdurl==0.1.2
mistralai==1.10.0
nexus-rpc==1.1.0
openai==2.15.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-ai==0.8.1
pydantic-ai-slim==0.8.1
pydantic-evals==0.8.1
pydantic-graph==0.8.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
temporalio==1.16.0
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
types-protobuf==6.32.1.20251210
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.39.0
wcwidth==0.2.14
websockets==15.0.1
zipp==3.23.0
PyMuPDF
reportlab
python-jose
passlib
bcrypt
python-multipart
fastapi-mail
apscheduler
firebase-admin
aiosmtplib
